➕[ADD] : alert Extension 추가
