+++ sopt30th_4th assignment /sopt30th_4th assignment/Source/Screen/Auth/Login/LoginViewController.swift
@@ -100,29 +100,12 @@ extension LoginViewController {
                 print(data)
             case .pathErr(let data):
                 guard let data = data as? LoginResponse else { return }
-                data.status == 404 ? self.alert(message: "해당하는 계정이 없습니다. ") : self.alert(message: "비밀번호를 다시 입력하세요.")
+                data.status == 404 ? self.alert(message: "존재하지 않는 계정입니다.\n다시 입력하세요. ") : self.alert(message: "비밀번호가 틀렸습니다. \n비밀번호를 다시 입력하세요.")
                 print(data.message)
             default:
                 self.alert(message: "로그인 실패")
             }
         }
-    }
-    
-    func alerttransition(message: String, storyboard: String, viewcontroller: String) {
-        let alertVC = UIAlertController(title: message, message: nil, preferredStyle: .alert)
-        let okAction = UIAlertAction(title: "확인", style: .default){_ in
-            let viewcontroller = UIStoryboard(name: storyboard, bundle: nil).instantiateViewController(withIdentifier: viewcontroller)
-            (UIApplication.shared.connectedScenes.first?.delegate as? SceneDelegate)?.changeRootVC(viewcontroller, animated: false)
-        }
-        alertVC.addAction(okAction)
-        present(alertVC, animated: true)
-    }
-
-    func alert(message: String) {
-        let alertVC = UIAlertController(title: message, message: nil, preferredStyle: .alert)
-        let okAction = UIAlertAction(title: "확인", style: .default, handler: nil)
-        alertVC.addAction(okAction)
-        present(alertVC, animated: true)
     }
 }
